fix(deploy, package): propagate non-zero exit codes
Because the err returned by `cmd.Wait()` wasn't checked `sam package` and `sam
deploy` would always return exit code 0 even if the aws cli cmd had a non-zero
exit code.

This commit checks the error and calls `os.Exit(1)`. This does not propagate the
original error code but exits with 1. This is because there seems to be no easy
os independent way of getting the exit code. Anyway this should do it for all
sane use cases.