test(event): add failing test for undefined PathParameters
As shown by this failing test sam-local returns an empty map if no path
parameters are defined on either the request or the route. Sadly this is not
the behaviour of API gateway. There it is null which causes untyped scripting
languages to work fine locally but crash on AWS.