chore: Update Flask version to 1.0.2
Functional tests for the local APIGW we intermittently failing due to
connection issues. This is due to the variable time periods the local
APIGW needs to get setup. Adding a sleep statement to the setup methods,
allows time for the setup to finish. Pervisouly, this was only in one tests
setup method.