fix: Options requests not invoking Local Lambda
Flask will automagically add a OPTIONS to a route if OPTIONS is not provided
in the list of HTTP Verbs on a url rule. Since we add a new url rule for each
path + HTTP Verb, Flask was adding OPTIONS to every path. This causes some
paths to not execute the Local Lambda code when calling a path with an OPTIONS
verb. We turn off this behavior by turning off the feature through provide_automatic_options.