fix: Tell Flask we are running the service in main (#491)
Flask thinks we are running in a production environment and displays
warning banners beause of it. We set an environment variable before
starting the service to tell Flask we are running within a 'main'
function.

This commit also adds a tests for having two Lambda functions defined under
a given path but have different methods associated with them.