Make nodejs8.10 template fully async/await
The current code mixes async/await with callbacks. This is unintuitive
and raises questions about which takes precedence: callback response
values or handler-returned promise resolved values.

The idiomatic way to write such a function is to rely on the return
value from the async handler. This cleans the code up substantially.