update installation.rst
this change offers info on the minimum system requirements for a redhat like system to get sam-cli up and running
this would be very similar for debian based system, but I haven't tested. If you feel it will be valuable I can do so.
the dockerfile is because "lets put everything in docker" I havent tried the lambda-local option in it. I know people run docker in docker but I would expect memory or file space issues
however if you want to have docker environment for plenty of other sam-cli related tasks this can be useful and it runs on  alpine so its small and fast.
I would probably guess not having this info on the main install page is good, I would think linking to it might be better, but you can tell me what is best.