fix: Support attaching to host network
Address #669

Docker only supports attaching containers to the host network through
the network_mode in the create. Previously, we tried to attach the
container to the host network after creating. This caused the command
to fail. Docker does not support attaching a container to host and bridge.
Therefore, we need to handle host as a special case.