fix(invoke/build): Allow AmazonLinux1803 Layer and fail when AmazonLinux1703 Layer is provided (#1181)
DockerLambda only supports one version of a container. Recently this was updated to AmazonLinux1803
but (at the time of writing) the Lambda Service is still using AmazonLinux1703. Given that, SAM CLI
will now fail if the arn:aws:lambda:::awslayer:AmazonLinux1703 Layer Arn is provided. If
arn:aws:lambda:::awslayer:AmazonLinux1803 Layer Arn is provided, we simply ignore it when reading
the Layer Definition on a Function. This scenarios happen in both build and invoke related commands.