Revert "Revert "fix: Respect zipped symlinks (#1140)""
This reverts commit 360d90aa560e0b2bb926ad037729ae160ebf17f6.

This commit originally reverted #1140 due to os.symlink not being
available on windows in Python2.7 stdlib, details here: https://github.com/awslabs/aws-sam-cli/pull/1315#issuecomment-518939441. We recently removed Python2.7 support in #1416, so this commit
is a revert of the revert, with some additional black formating to make `make pr`
pass.