chore: fix java build and ruby appveyor builds
Split java and ruby builds into two functions, one for in process and one for container.
When running java builds on Windows Server 2019, the in process builds where hanging,
reason is known. Splitting the tests allows us to run the container builds on
Windows Server 2019, which is required due to needing docker that spins up a linux
instance. We can then run in-process builds in Visual Studio images, which works for java.

This commit will also update the ruby builds on windows but instead of running both in process
and container builds, we will only run container builds. When running ruby in-process builds
on Visual Studio images, we still see failures (which is assumed to be the same as when they
are run on Windows Server 2019 which is an ssl issue). This will at least allow us to verify
container builds for ruby, which will give us confidence that things are working as expected.