Change How App Templates Are Updated for `sam init` (#1678)
Addresses several issues around IDE init as well as Windows init. With this change, the `sam init` command will clone the app templates repo automatically (with shell notifications), and uses a two-step update process so that the existing app templates copy on a user's machine is not deleted unless we have already successfully downloaded a replacement.