Add an integration test on OPTIONS handling
https://github.com/awslabs/aws-sam-cli/issues/1434

This adds an integration test to verify that templates with explicit
OPTIONS handlers actually have those handlers invoked on OPTIONS
requests. There has been an issue with OPTIONS requests returning 200
and not reaching an explicitly specified OPTIONS handler.
https://github.com/awslabs/aws-sam-cli/pull/1649 fixes that, so if the
integration test added in this commit passes (and fails on the develop
branch), it could be added to that PR.