Code Sign Integration  (#217)
* Release v1.0.0 (#2111)

* feat: Use aws-sam-cli docker images (#2066)

* Add Source for Docker Build Images (#2078)

* chore: Bump AWS SAM CLI Version (#2079)

* Version bump (#2080)

* chore: Bump AWS SAM CLI Version

* Change SAM CLI Version Number

There is a conflict betweeen PyPi documentation which asks for the
previous style
https://packaging.python.org/guides/distributing-packages-using-setuptools/#pre-release-versioning
and PEP 440 which proposes the style included in this change
https://www.python.org/dev/peps/pep-0440/#pre-releases - our MSI build
scripts failed on the pattern we were using before, this changes the
pattern.

* refactor: Build init.go with -s and -w flags to removed debug info (#2083)

* refactor: Bake Rapid into image on the fly (#2100)

* refactor: Bake Rapid into image on the fly

* force chmod on init binary in container for windows

* bake go debug bootstrap

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>

* chore: Bump version to RC2 (#2104)

* Remove liblzma and libxslt from AL2 build images (#2109)

Discovered a regression where on Ruby 2.7, the `nokogiri` dependency
would build without errors, but would not run on local testing or on AWS
Lambda itself.

On further investigation, it appears that `nokogiri` can compile with or
without `liblzma` present, but if it is present in the build
enviornment (pre-change) and it is not present on the invoke
environment (true in AL2 runtimes), you will experience runtime failures
attempting to require `nokogiri`.

I have been able to verify that with these changes, `nokogiri` builds
correctly for Ruby 2.7 and runs both locally and on AWS Lambda.

* Build output dots (#2112)

* Use Low-Level Docker Client

Allows us to stream dots as a progress heartbeat. Pending unit tests and
black formatting.

* Get make pr Passing

Co-authored-by: Jacob Fuss <32497805+jfuss@users.noreply.github.com>
Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>

* chore: Bump aws-lambda-builders and SAM CLI to 1.0.0 (#2116)

* fix: Update Python3.8 debug entrypoint (#2119)

* chore: readme update with screenshot (#2117)

* chore: readme update with screenshot

* chore: remove beta in the title

Co-authored-by: Alex Wood <awood45@gmail.com>

* feature: Lambda Code Sign integration for SAM CLI

* feature: Lambda Code Sign integration for SAM CLI (actual signing impl and unit tests)

* Add details to print_deploy_args
Add documentation for missing classes and methods

* Update couple of prompts

* Wording changes requested by UX & Docs Team

Co-authored-by: Alex Wood <awood45@gmail.com>
Co-authored-by: Jacob Fuss <32497805+jfuss@users.noreply.github.com>
Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>
Co-authored-by: Sriram Madapusi Vasudevan <3770774+sriram-mv@users.noreply.github.com>