Fix schema loading when no default profile
The region_name value when a profile does not specify a region or when
there is not a 'default' profile resulted in a crash since
_get_aws_region_choice attempted to perform concatenation using the `+`
operator (which is undefined for str + NoneType). Using an actual string
formatting method fixes that issue.

Additionally, it is useless to prompt the user whether not to use the
'default' profile since doing so is impossible. They need to be
prompted. Additionally, it is slightly more 'intelligent' to create a
new session after the user has chosen a new profile so that a better
default region can be chosen.