fix: Add all files under samcli module
When we released the `sam pipeline` set of commands, we didn't update
the MANIFEST.in file. This file controls what files get added or ignored
when the .whl file is created. When we add non python files, we need to
remember to update the MaNIFEST.in file. This commit will auto include
any files within the samcli module but exlcude .pyc files. This might
create a bigger .whl but will make sure the cli works going forward.

I did some quick tests and compared 1.28 .whl to one I produced with this
commit. I did not see any differences in the diff.