fix: Add all files under samcli module (#3177)
* fix: Add all files under samcli module

When we released the `sam pipeline` set of commands, we didn't update
the MANIFEST.in file. This file controls what files get added or ignored
when the .whl file is created. When we add non python files, we need to
remember to update the MaNIFEST.in file. This commit will auto include
any files within the samcli module but exlcude .pyc files. This might
create a bigger .whl but will make sure the cli works going forward.

I did some quick tests and compared 1.28 .whl to one I produced with this
commit. I did not see any differences in the diff.

* Sleep for 1s in test_must_record_function_duration to help avoid flaky test runs

* Update tests/unit/lib/telemetry/test_metric.py

Update comment from ms to s

Co-authored-by: Cosh_ <CoshUS@users.noreply.github.com>

* Update tests/unit/lib/telemetry/test_metric.py

Co-authored-by: Cosh_ <CoshUS@users.noreply.github.com>

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>
Co-authored-by: Cosh_ <CoshUS@users.noreply.github.com>
Co-authored-by: Raymond Wang <14915548+wchengru@users.noreply.github.com>