updates the local API service to include CORS headers
Previous to this commit, `sam local api` would only return CORS headers
when an OPTIONS request is made. This causes issues for browsers, which
expect at the very least "Access-Control-Allow-Origin" to be present on
other HTTP methods. This commit adds the "Access-Control-Allow-Origin"
header to a request if it is made with one of the Allowed-Methods.

References:
 * Fetch spec on CORS: https://fetch.spec.whatwg.org/#cors-request
 * Examples from Mozilla: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS