Fix Lambda invoke in user networks
The Docker CLI client won't let you attach containers to a user
network and the bridge at the same time, presumably for a good
reason. One good reason seems to be that name resolution doesn't
work - which means that if you pass the service name of e.g.
`localstack` on the same user network to provide endpoints for AWS
APIs for testing, it's not going to work. Disconnecting the container
from the bridge network seems to fix this.

Secondly, if SAM is running on a container inside the same user
network (the likeliest scenario for using one?) the Lambda ports are
not available on `localhost`, and the remapped ports are not available
inside the network.

Therefore if the container is in a user network :

- Detach it from the bridge network
- Set its hostname to the first 12 chars of the container ID
- Set its rapid port to 8080

Then invocations work properly.