Make validation errors fatal
Up until now, running `sam validate` on an invalid template did print
warning to the console, but the last message printed by the sam
command was still claiming that the template is valid.

This happens because the template translation process that is used as
the validation provider does not consider validtion errors to be fatal
and ignores them.

This commit makes the validation errors fatal by running validation in
a separate step and terminate the validation process if the template
is not considered valid by the validator.