fix: multivalue query params http (#3788)
* fixes multivalue query param requests for HTTP APIs

When a query parameter is specified multiple times in a URL, e.g.
`path?mykey=1&mykey=2&mykey=3`
this should create an event with all appropriate values.

Prior to this commit, when using `sam local start-api` with an HTTP API,
only the first value is present in the queryStringParameters of the
event. In payload format v2 [1] used in HTTP APIs, multiple values
should be combined with commas and included in the queryStringParameters
field. This commit fixes the behavior for query string parameters.

[1] https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html

* black reformat

* apply suggestion

* add test

Co-authored-by: Jason LaPier <jason@heyfieldday.com>
Co-authored-by: Qingchuan Ma <69653965+qingchm@users.noreply.github.com>
Co-authored-by: Mohamed Elasmar <71043312+moelasmar@users.noreply.github.com>
Co-authored-by: Jacob Fuss <32497805+jfuss@users.noreply.github.com>
Co-authored-by: Mehmet Nuri Deveci <5735811+mndeveci@users.noreply.github.com>