Support pipeline templates manifests in custom Git locations
When a custom Git location is passed to `sam pipeline init`, the utility
will first look for a valid pipeline template manifest in the repository's
root. If one is found, the user will be prompted to choose which pipeline
to create. If one is not found, the repository will be treated as a
standard pipeline template.