Fix failing test
This test invokes the `InteractiveInitFlow` and passes in a local
templates path whose files are mocked and thus don't actually exist.
As a result, the test would fail when `_generate_from_custom_location()`
would attempt to operate on a template manifest file which was not
readable. This change uses a mock side effect to merely returns a
`False` value for when `os.path.exists()` is called for the template
manifest which prevents the failure condition from occurring.