fix: test_custom_build_image_succeeds build test was failing due log not matching (#5030)
* fix: test_custom_build_image_succeeds build test was failing due log not matching

In #5016, we split the image_name and tag to allow for better support with other
docker like systems (podman). The side affect was that, we logged the "image:tag"
and asserted that it would be in the logs. Due to the change, the log was changed
and caused the test to fail.

* Fix unit test

---------

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>