fix(list): Handle Stacks with >100 resources
Previously, we were using `describe_stack_resources` to get all the
Resources from a Stack. However, this will only ever return the
first 100 resources (per API docs). Instead, we move to using the
`list_stack_resources` paginator (per API docs). This allows us to
get all the resources within a stack. The response is different from
`describe_stack_resources` and `list_stack_resources` but the difference
is not meaningful for the current implementation.

As apart of this, I updated the unit tests. Previously, we were using another
class to get a CFN Client to mock. Instead of mocking out a file that is not
used within the `resource_mapping_producer.py`, I remove that and mocked out
the `cfn_client` directly. This makes the tests easier to follow and removes
unneeded mocking of classes/methods that are not used within the implementation.