chore: GHA for Auto Closing issues that were released
This GHA closes any issue that is marked "stage/waiting-for-release".
This may close issues that are linked to PRs that were merged into develop
but not yet released. There was no easy way to get all issues that
are linked to a PR closed between two commits. Instead of doing long queries
to get the exact issues, we will rely on the labels being updated and
handle false positives as/if they happen.