chore: Upgrade Linux installer to Python3.11 from 3.7 (#5156)
* chore: Upgrade Linux installer to Python3.11 from 3.7

Python3.7 is reaching end of life and is starting to cause issues
with upgrading dependencies, since newer versions are dropping support.

This moves the linux installer to the latest Python3.11 which puts
us into a much better state for dependencies and allows us to be
more current with our Python versions in general.

This only covers linux, we should match all our installers onto the same
Python version but linux is the only one on Python3.7 right now, so more urgent
to upgrade.

* Adjust order of inputs

* Make a reproducible-mac to make mac pyinstaller tests pass in GHA

* Add make command to update both reproducible files

---------

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>