feat: sam Commands Understand Local File Paths for `AWS::Serverless::Function.ImageUri` (#6930)
* feat: sam Commands Understand Local File Paths for `ImageUri`

As a summary, sam has learned to load an an image from a local archive
before proceeding with the `build`, `package`, and `deploy` commands.

When running `sam build` with an `ImageUri` pointing to a local file,
sam will load that archive into an image, then write the ID of the
image to the `ImageUri` property of the built template. ID works the
same as a tag for the Docker API, so business continues as usual from
here. The reason behind writing ID is that a loaded image could be
associated with multiple tags, and selecting one arbtrarily leads to
difficulties in the deploy command.

The package and deploy commands have three kinds of value for
`ImageUri` to consider. First, a value of the form
`{repo}:{tag}`. This functions as it always has. Second, an image
ID (in the form `sha256:{digest}`) which is probably the output of
`sam build`. In this case, the tag translation uses the name of the
as its input. Otherwise, they'd all end up with names starting with
"sha256". Last, a local file. In this case, it proceeds as it does in
the build command: Load the archive into an image first, then pass the
resource name into tag translation.

See: #6909

* Reword Explanatory Comment

See: #6909

* Correct Old Typo in String Parameter

See: #6909

* Take a Swing at Unit Tests

See: #6909

* Cover Another Test Case

See: #6909

* Take a Swing at Integration Tests

Also, genericize unit tests a little.

See: #6909

* Now Add Package Integration Tests

* And Deploy Integration Tests

* Point to Correct File, Remove Unused Import

---------

Co-authored-by: jysheng123 <141280295+jysheng123@users.noreply.github.com>
Co-authored-by: sidhujus <105385029+sidhujus@users.noreply.github.com>