feat: Support asynchronous shareable test events (#8373)
This feature was added recently to the Lambda console. It works by
storing event metadata as an `x-metadata` OpenAPI extension on the
schema's example.

Here we support putting an event with a desired `--invocation-type`
and passing that to a remote invoke (with an explicitly provided
invocation type `--parameter` taking precedence).

Co-authored-by: seshubaws <116689586+seshubaws@users.noreply.github.com>